requirements.txt
absl-py==2.0.0
accelerate
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
apptools==5.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.3
blessed==1.20.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
clarabel==0.6.0
click==8.1.7
configobj==5.0.8
contourpy==1.1.1
cvxpy==1.4.1
cycler==0.12.1
datasets==2.15.0
dill==0.3.7
easydict==1.11
ecos==2.0.12
einops==0.6.0
envisage==7.0.3
exceptiongroup==1.2.2
fastapi==0.115.2
ffmpy==0.3.2
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
gast==0.4.0
gdown==5.1.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.59.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.19.4
hydra-core==1.3.2
idna==3.4
imageio==2.34.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
ipdb==0.13.13
jinja2==3.1.2
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
markdown==3.5
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
orjson==3.10.7
osqp==0.6.3
pandas==2.0.3
pillow==9.5.0
portalocker==2.10.1
prompt-toolkit==3.0.41
protobuf==4.24.4
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pydantic==2.9.2
pydantic-core==2.23.4
pydub==0.25.1
pyface==8.0.0
pyglm==2.7.0
pyparsing==3.1.1
pyqt5==5.15.10
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pysocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.12
pytz==2023.3.post1
pyyaml==6.0
qdldl==0.1.7.post0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.2
rsa==4.9
ruff==0.6.9
safetensors==0.4.1
scipy==1.9.3
scs==3.2.3
semantic-version==2.10.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.3
termcolor==2.4.0
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.1
torchaudio
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.0
traits==6.4.3
traitsui==8.0.0
transformers==4.35.2
triton==3.0.0
typer==0.12.5
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.32.0
vtk==9.3.0
wcwidth==0.2.12
webdataset==0.2.96
websockets==12.0
werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
